pyproject.toml
[project]
name = "ai-devs4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
]
